Update Etcher.download.recipe
It seems that they're pushing on with changes after the Resin.io rebrand and they're now altering the name of the download DMG and app bundle. I've tested these changes locally and they seem to now download the latest version (your current recipe seems to grab 1.4.4) but please test as I'm pretty new to AutoPKG recipes.