Update Etcher.download.recipe
With the recent rebranding from Etcher to balenaEtcher, there are no current releases (only pre-releases) with the new name format. Until a release happens, this download recipe is broken with the current regex. Either the regex could be updated, or the proposed changes could be put in place and then removed after the first release.

`Error in local.munki.autopkg_Etcher: Processor: GitHubReleasesInfoProvider: Error: No release assets were found that satisfy the criteria.`